[API-92] Extract the placement engine into placement.ts.
