+++ api/schedules/dynamic/placement.ts
@@ -1,0 +1,320 @@
+import type dayjs from 'dayjs';
+import type { Zone, IrrigationScheduleEntry, IrrigationCycle } from '@/models';
+import type { SchedulingDecisionReason } from '@/models/decision';
+import { isDayAllowed, isNightSkipped, type ScheduleRestrictions } from './restrictions';
+import { roundTo1Decimal } from './util';
+import type { BusyWindow } from '.';
+
+type PlaceIrrigationInputs = {
+    date: dayjs.Dayjs;
+    sunrise: dayjs.Dayjs;
+    zone: Zone;
+    currentDepletionMillimeters: number;
+    totalAvailableWaterMillimeters: number;
+    precipitationRateMillimetersPerHour: number;
+    infiltrationRateMillimetersPerHour: number;
+    soakTimeMinutes: number;
+    busyWindowsSoFar: ReadonlyArray<BusyWindow>;
+    restrictions: ScheduleRestrictions;
+};
+
+/**
+ * Result of `tryPlaceIrrigationForDay`. Either cycles were `placed` (with the
+ * resulting entry and a `partial` flag set when the overnight window clamped
+ * the refill, API-75), or the day was `skipped` / `deferred` with the specific
+ * reason — the caller persists that reason into the decisions log (API-88) and
+ * carries depletion forward unchanged.
+ */
+export type PlaceIrrigationResult =
+    | { kind: 'placed'; cycles: IrrigationCycle[]; entry: IrrigationScheduleEntry; partial: boolean }
+    | { kind: 'skipped' | 'deferred'; reason: SchedulingDecisionReason };
+
+/**
+ * Computes the day's irrigation block, anchored so the last cycle ends at
+ * the next day's sunrise and cycles fill backward into the night. Returns a
+ * `skipped` / `deferred` result when the day is disallowed or no cycles fit the
+ * overnight window — callers treat that as "skip, carry depletion forward."
+ *
+ * The overnight window is [midnight of cycle day, sunrise of cycle day] —
+ * derived from `sunrise.startOf('day')` and `sunrise`. With the API-76 anchor
+ * shift, the cycle day is the day *after* the planning day, so for an evening
+ * replan on Sun the block runs [Mon 00:00, Mon sunrise] — naturally in the
+ * future.
+ *
+ * Cycle sizing uses an iterative downward retry (API-77): the closed-form
+ * `maxRunTimeMinutes` gives an upper bound on N against the empty window,
+ * but cross-zone busy windows can consume so much of the residual gaps that
+ * the placer can't fit N cycles. The caller retries with N-1, N-2, … until
+ * placement succeeds or N=0. The `actualTotalRunTime` drives the partial-
+ * refill `appliedDepthMm` / `depletionAfterMm` math.
+ *
+ * Past-window busy intervals (those starting at the epoch) are split out
+ * from cross-zone busy windows. Once the placer has produced the cycle list,
+ * any cycle whose `startTime` falls before `pastWindow.end` (= `now`) is
+ * dropped. In the normal evening-replan flow no cycle is dropped; the filter
+ * only bites for in-flight replans (`now` lies inside the planning block).
+ */
+export function tryPlaceIrrigationForDay(inputs: PlaceIrrigationInputs): PlaceIrrigationResult {
+    const {
+        date, sunrise, zone, currentDepletionMillimeters, totalAvailableWaterMillimeters,
+        precipitationRateMillimetersPerHour, infiltrationRateMillimetersPerHour, soakTimeMinutes,
+        busyWindowsSoFar, restrictions,
+    } = inputs;
+
+    if (!isDayAllowed(restrictions, date.isoWeekday())) {
+        console.warn(`planner: zone ${zone.id} (${zone.name}): day ${date.format('YYYY-MM-DD')} (isoWeekday ${date.isoWeekday()}) disallowed by schedule restrictions — skipping irrigation.`);
+        return { kind: 'skipped', reason: 'day-not-allowed' };
+    }
+
+    if (isNightSkipped(restrictions, date)) {
+        console.warn(`planner: zone ${zone.id} (${zone.name}): day ${date.format('YYYY-MM-DD')} is skip-marked — skipping irrigation.`);
+        return { kind: 'skipped', reason: 'operator-skip' };
+    }
+
+    const depletionBeforeIrrigation = currentDepletionMillimeters;
+    const fullRefillGrossDepthMillimeters = Math.min(
+        currentDepletionMillimeters / zone.irrigationEfficiency,
+        totalAvailableWaterMillimeters,
+    );
+    const fullRefillRunTimeMinutes = (fullRefillGrossDepthMillimeters / precipitationRateMillimetersPerHour) * 60;
+    const maximumCycleMinutes = infiltrationRateMillimetersPerHour > 0
+        ? (infiltrationRateMillimetersPerHour / precipitationRateMillimetersPerHour) * 60
+        : fullRefillRunTimeMinutes;
+
+    // Overnight floor: midnight (00:00 local) of the irrigation entry's date.
+    // No cycle may start before this time.
+    const earliestStart = sunrise.startOf('day');
+
+    // Cap runtime at what tonight's overnight window can physically hold
+    // (API-75). The closed-form picks the largest N where
+    // `N·maxCycle + (N-1)·soak ≤ windowMinutes`, then maxRunTime = N·maxCycle.
+    // When the full-refill runtime exceeds this cap, we accept a partial
+    // refill — depletionAfter reflects the residual and carries forward to
+    // the next allowed day. Without this clamp, deep deficits on low-
+    // infiltration soils defer forever because every full refill is longer
+    // than [midnight, sunrise].
+    const windowMinutes = sunrise.diff(earliestStart, 'minute');
+    const maxRunTimeMinutes = computeMaxRunTimeMinutes(maximumCycleMinutes, soakTimeMinutes, windowMinutes);
+    if (maxRunTimeMinutes <= 0) {
+        console.warn(`planner: zone ${zone.id} (${zone.name}): overnight window (${windowMinutes} min) too short to fit even one cycle (maxCycle ${maximumCycleMinutes.toFixed(1)} min) on ${date.format('YYYY-MM-DD')} — skipping irrigation.`);
+        return { kind: 'deferred', reason: 'window-too-short' };
+    }
+    const initialTotalRunTimeMinutes = Math.min(fullRefillRunTimeMinutes, maxRunTimeMinutes);
+
+    // The daemon plumbs a "past window" — { start: epoch, end: now } — as a
+    // busy interval that pushes past-dated cycles forward. It's handled
+    // separately from real cross-zone busy windows: cross-zone windows feed
+    // the backward placer (so cycles slide *earlier* into soak gaps), while
+    // the past window applies as a forward shift *after* placement so a
+    // past-due cycle is fired now-ish instead of dropped.
+    const pastWindow = busyWindowsSoFar.find(w => w.start.valueOf() === 0);
+    const crossZoneBusyWindows = pastWindow
+        ? busyWindowsSoFar.filter(w => w !== pastWindow)
+        : busyWindowsSoFar;
+
+    // Iterative downward cap (API-77). The closed-form `maxRunTimeMinutes`
+    // sizes against the *empty* overnight window; with earlier zones already
+    // planted, the residual gaps may not fit the chosen N cycles. Try N, then
+    // N-1, etc., until the placer succeeds. Shrinking N is more honest than
+    // a residual-time closed form because interleaving into soak gaps fits
+    // cycles when arithmetic says no. The first non-null placement wins.
+    const initialN = computeInitialCycleCount(initialTotalRunTimeMinutes, maximumCycleMinutes);
+    let placedCycles: IrrigationCycle[] | null = null;
+    let actualTotalRunTimeMinutes = 0;
+    for (let n = initialN; n >= 1; n--) {
+        const candidateRunTime = candidateTotalRunTime(n, fullRefillRunTimeMinutes, maximumCycleMinutes);
+        const candidatePerCycle = roundTo1Decimal(candidateRunTime / n);
+        if (candidatePerCycle <= 0) continue;
+        const candidate = placeCyclesBackwardAvoidingBusy(
+            n,
+            candidatePerCycle,
+            sunrise,
+            soakTimeMinutes,
+            earliestStart,
+            crossZoneBusyWindows,
+        );
+        if (candidate !== null) {
+            placedCycles = candidate;
+            actualTotalRunTimeMinutes = n * candidatePerCycle;
+            break;
+        }
+    }
+
+    if (placedCycles === null) {
+        console.warn(`planner: zone ${zone.id} (${zone.name}): no cycle count from ${initialN} down fit the overnight window's residual gaps on ${date.format('YYYY-MM-DD')} — deferring irrigation.`);
+        return { kind: 'deferred', reason: 'no-cycle-fit' };
+    }
+
+    if (placedCycles.length === 0) {
+        console.warn(`planner: zone ${zone.id} (${zone.name}): no cycles fit the overnight window on ${date.format('YYYY-MM-DD')} — skipping irrigation.`);
+        return { kind: 'deferred', reason: 'no-cycle-fit' };
+    }
+
+    const grossIrrigationDepthMillimeters = (actualTotalRunTimeMinutes / 60) * precipitationRateMillimetersPerHour;
+    if (actualTotalRunTimeMinutes < fullRefillRunTimeMinutes) {
+        console.log(`planner: zone ${zone.id} (${zone.name}): clamping run time ${fullRefillRunTimeMinutes.toFixed(1)} → ${actualTotalRunTimeMinutes.toFixed(1)} min on ${date.format('YYYY-MM-DD')} (window ${windowMinutes} min, ${placedCycles.length} cycle(s)); applying ${grossIrrigationDepthMillimeters.toFixed(1)} of ${fullRefillGrossDepthMillimeters.toFixed(1)} mm.`);
+    }
+
+    // Past-window handling (API-76). The block now anchors to the next day's
+    // sunrise, so for an evening replan the placed cycles already sit in the
+    // future. The only case where the past window matters is the in-flight
+    // replan — `now` lands inside [midnight cycle-day, sunrise cycle-day] —
+    // and we drop any cycle whose start has already passed. No forward shift,
+    // no midnight filter, no endBySunrise drop: the new anchor already
+    // guarantees `cycleEnd ≤ sunrise` and that all surviving cycles are
+    // overnight by construction.
+    const finalCycles = pastWindow === undefined
+        ? placedCycles
+        : placedCycles.filter(c => !c.startTime.isBefore(pastWindow.end));
+
+    if (finalCycles.length === 0) {
+        console.warn(`planner: zone ${zone.id} (${zone.name}): no cycles remain after past-window handling on ${date.format('YYYY-MM-DD')} — skipping irrigation.`);
+        return { kind: 'deferred', reason: 'past-window' };
+    }
+
+    // Net depth that actually reaches the root zone after losses. `depletionAfter`
+    // is the residual depletion after this event — zero on a full refill,
+    // positive when the overnight-window clamp limited the applied gross
+    // (API-75). The planner's day loop carries this residual into tomorrow.
+    const netAppliedDepthMillimeters = grossIrrigationDepthMillimeters * zone.irrigationEfficiency;
+    const depletionAfterIrrigation = Math.max(0, depletionBeforeIrrigation - netAppliedDepthMillimeters);
+
+    const entry: IrrigationScheduleEntry = {
+        date,
+        zoneId: zone.id,
+        cycles: finalCycles,
+        appliedDepthMm: roundTo1Decimal(grossIrrigationDepthMillimeters),
+        depletionBeforeMm: roundTo1Decimal(depletionBeforeIrrigation),
+        depletionAfterMm: roundTo1Decimal(depletionAfterIrrigation),
+        sunriseAt: sunrise,
+    };
+    // `partial` when the overnight window clamped the applied gross below a
+    // full refill (API-75) — distinguishes `partial-refill` from `full-refill`
+    // in the decisions log (API-88).
+    return { kind: 'placed', cycles: finalCycles, entry, partial: actualTotalRunTimeMinutes < fullRefillRunTimeMinutes };
+}
+
+/**
+ * Returns the largest total runtime (minutes) whose cycle layout fits inside
+ * `windowMinutes` given the per-cycle infiltration cap (`maxCycleMinutes`) and
+ * the intra-zone soak gap. Closed form: pick the largest `N` where
+ * `N·maxCycle + (N-1)·soak ≤ windowMinutes`, then `maxRunTime = N·maxCycle`.
+ * Returns 0 when not even one minimum-width cycle fits.
+ *
+ * For `maxCycleMinutes <= 0` (no infiltration limit), a single uninterrupted
+ * cycle is allowed up to the full window — there's no per-cycle ceiling.
+ *
+ * Sizes against the *empty* window — API-77 wraps this in an outer downward
+ * retry that shrinks N when the residual gaps after cross-zone placement
+ * can't hold the closed-form plan.
+ */
+function computeMaxRunTimeMinutes(maxCycleMinutes: number, soakTimeMinutes: number, windowMinutes: number): number {
+    if (windowMinutes <= 0) return 0;
+    if (maxCycleMinutes <= 0) return windowMinutes;
+    const cyclesThatFit = Math.floor((windowMinutes + soakTimeMinutes) / (maxCycleMinutes + soakTimeMinutes));
+    if (cyclesThatFit <= 0) return 0;
+    return cyclesThatFit * maxCycleMinutes;
+}
+
+/**
+ * Returns the initial cycle count (N) the iterative downward retry should
+ * start from, given a total runtime and per-cycle infiltration cap. Matches
+ * the pre-API-77 placer's internal derivation: N=1 for the no-cap / single-
+ * cycle case, otherwise `ceil(totalRunTime / maxCycle)`.
+ */
+function computeInitialCycleCount(totalRunTimeMinutes: number, maximumCycleMinutes: number): number {
+    if (totalRunTimeMinutes <= 0) return 0;
+    if (maximumCycleMinutes <= 0 || totalRunTimeMinutes <= maximumCycleMinutes) return 1;
+    return Math.ceil(totalRunTimeMinutes / maximumCycleMinutes);
+}
+
+/**
+ * Returns the candidate total runtime for `n` cycles inside the iterative
+ * downward retry: capped at the full-refill runtime (no point watering past
+ * a full refill) and at `n * maxCycle` (the infiltration ceiling for `n`
+ * cycles). When `maxCycle <= 0` (no infiltration limit) the only cap is
+ * the full refill.
+ */
+function candidateTotalRunTime(n: number, fullRefillRunTimeMinutes: number, maximumCycleMinutes: number): number {
+    if (maximumCycleMinutes <= 0) return fullRefillRunTimeMinutes;
+    return Math.min(fullRefillRunTimeMinutes, n * maximumCycleMinutes);
+}
+
+/**
+ * Builds and places an irrigation cycle plan anchored so the last cycle ends
+ * at sunrise, walking backward into the night. While walking, each cycle is
+ * slid *earlier* past any overlapping `busyWindows` until it fits — that's
+ * what lets a follow-on zone's cycles interleave into the soak gaps of an
+ * earlier-planned zone instead of being shoved past sunrise (the old
+ * forward-only deconflict cascaded delays and dropped most days for the
+ * second zone, see API-66).
+ *
+ * Interleaving invariant (API-73). On overlap the cycle snaps to
+ * `cycleEnd = busy.start`, so a follow-on zone's cycle ends flush against the
+ * earlier zone's next start — a 0-minute inter-zone gap. Intra-zone soak is
+ * preserved via `cursor`, which walks backward by `soakTimeMinutes` from each
+ * placed cycle's start. Together these two rules cause a follow-on zone's
+ * cycles to land inside the earlier zone's soak gaps whenever they
+ * dimensionally fit, compressing the multi-zone overnight block.
+ *
+ * When a cycle still doesn't fit (sliding earlier would land before
+ * `earliestStart` = midnight), the entire day is treated as un-irrigatable:
+ * the function returns `null` so the caller defers and carries depletion
+ * into the next allowed day. The caller (API-77) wraps this in an iterative
+ * downward retry, calling with progressively smaller `numberOfCycles` until
+ * placement succeeds.
+ *
+ * @param numberOfCycles - Number of cycles to lay down (chosen by the caller's
+ *   downward retry). Returns `[]` when zero.
+ * @param perCycleMinutes - Per-cycle duration. The caller picks this such
+ *   that `numberOfCycles * perCycleMinutes` matches the desired total runtime.
+ * @param sunrise - End anchor: last cycle ends here (or earlier if busy windows displace it).
+ * @param soakTimeMinutes - Soak gap between consecutive cycles of this zone.
+ * @param earliestStart - Hard floor: no cycle may start before this time (midnight of the entry date).
+ * @param busyWindows - Intervals occupied by other zones (or the past window
+ *   if handled here). Cycles slide earlier to avoid overlap.
+ * @returns Cycles in chronological order (earliest first), `null` when the
+ *   day must be deferred, or `[]` when `numberOfCycles === 0`.
+ */
+function placeCyclesBackwardAvoidingBusy(
+    numberOfCycles: number,
+    perCycleMinutes: number,
+    sunrise: dayjs.Dayjs,
+    soakTimeMinutes: number,
+    earliestStart: dayjs.Dayjs,
+    busyWindows: ReadonlyArray<BusyWindow>,
+): IrrigationCycle[] | null {
+    if (numberOfCycles <= 0 || perCycleMinutes <= 0) return [];
+
+    const cyclesInReverse: IrrigationCycle[] = [];
+    let cursor: dayjs.Dayjs = sunrise; // latest-allowed end for the next placed cycle (walking backward)
+
+    for (let i = 0; i < numberOfCycles; i++) {
+        let cycleEnd: dayjs.Dayjs = cursor;
+        let cycleStart: dayjs.Dayjs = cycleEnd.subtract(perCycleMinutes, 'minute');
+
+        // Slide earlier past any busy window we overlap. Re-iterate after each
+        // shift since the new (earlier) slot may overlap a different window.
+        let stable = false;
+        while (!stable) {
+            stable = true;
+            for (const w of busyWindows) {
+                if (cycleStart.isBefore(w.end) && cycleEnd.isAfter(w.start)) {
+                    cycleEnd = w.start;
+                    cycleStart = cycleEnd.subtract(perCycleMinutes, 'minute');
+                    stable = false;
+                    break;
+                }
+            }
+        }
+
+        if (cycleStart.isBefore(earliestStart)) {
+            return null; // defer: window too short to fit the required cycles
+        }
+
+        cyclesInReverse.push({ startTime: cycleStart, durationMin: perCycleMinutes });
+        cursor = cycleStart.subtract(soakTimeMinutes, 'minute');
+    }
+
+    return cyclesInReverse.reverse();
+}
